# api/v1/types.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package v1

// Defines values for AgentModeRequestMode.
const (
	AgentModeRequestModeConnected    AgentModeRequestMode = "connected"
	AgentModeRequestModeDisconnected AgentModeRequestMode = "disconnected"
)

// Defines values for AgentStatusConsoleConnection.
const (
	AgentStatusConsoleConnectionConnected    AgentStatusConsoleConnection = "connected"
	AgentStatusConsoleConnectionConnecting   AgentStatusConsoleConnection = "connecting"
	AgentStatusConsoleConnectionDisconnected AgentStatusConsoleConnection = "disconnected"
	AgentStatusConsoleConnectionError        AgentStatusConsoleConnection = "error"
)

// Defines values for AgentStatusMode.
const (
	AgentStatusModeConnected    AgentStatusMode = "connected"
	AgentStatusModeDisconnected AgentStatusMode = "disconnected"
)

// Defines values for CollectorStatusStatus.
const (
	CollectorStatusStatusConnected           CollectorStatusStatus = "connected"
	CollectorStatusStatusConnecting          CollectorStatusStatus = "connecting"
	CollectorStatusStatusError               CollectorStatusStatus = "error"
	CollectorStatusStatusIdle                CollectorStatusStatus = "idle"
	CollectorStatusStatusInventoryCollected  CollectorStatusStatus = "inventory_collected"
	CollectorStatusStatusInventoryCollecting CollectorStatusStatus = "inventory_collecting"
)

// AgentModeRequest defines model for AgentModeRequest.
type AgentModeRequest struct {
	Mode AgentModeRequestMode `json:"mode"`
}

// AgentModeRequestMode defines model for AgentModeRequest.Mode.
type AgentModeRequestMode string

// AgentStatus defines model for AgentStatus.
type AgentStatus struct {
	// ConsoleConnection Current console connection status
	ConsoleConnection AgentStatusConsoleConnection `json:"console_connection"`

	// Mode Target mode for the agent
	Mode AgentStatusMode `json:"mode"`
}

// AgentStatusConsoleConnection Current console connection status
type AgentStatusConsoleConnection string

// AgentStatusMode Target mode for the agent
type AgentStatusMode string

// CollectorStartRequest defines model for CollectorStartRequest.
type CollectorStartRequest struct {
	Password string `json:"password"`

	// Url vCenter URL
	Url      string `json:"url"`
	Username string `json:"username"`
}

// CollectorStatus defines model for CollectorStatus.
type CollectorStatus struct {
	// Error Error message when status is error
	Error  *string               `json:"error,omitempty"`
	Status CollectorStatusStatus `json:"status"`
}

// CollectorStatusStatus defines model for CollectorStatus.Status.
type CollectorStatusStatus string

// SetAgentModeJSONRequestBody defines body for SetAgentMode for application/json ContentType.
type SetAgentModeJSONRequestBody = AgentModeRequest

// StartCollectorJSONRequestBody defines body for StartCollector for application/json ContentType.
type StartCollectorJSONRequestBody = CollectorStartRequest