# internal/handlers/console.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"

	v1 "github.com/tupyy/assisted-migration-agent/api/v1"
	"github.com/tupyy/assisted-migration-agent/internal/models"
)

// GetAgentStatus returns the current agent status
// (GET /agent)
func (h *Handler) GetAgentStatus(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"msg": "not implemented"})
}

// SetAgentMode changes the agent mode
// (POST /agent)
func (h *Handler) SetAgentMode(c *gin.Context) {
	var req v1.AgentModeRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}

	var mode models.AgentMode
	switch req.Mode {
	case v1.AgentModeRequestModeConnected:
		mode = models.AgentModeConnected
	case v1.AgentModeRequestModeDisconnected:
		mode = models.AgentModeDisconnected
	default:
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid mode: must be 'connected' or 'disconnected'"})
		return
	}

	h.consoleSrv.SetMode(mode)

	status := h.consoleSrv.Status()
	var resp v1.AgentStatus
	resp.FromModel(models.AgentStatus{Console: status})

	c.JSON(http.StatusOK, resp)
}